feat: 영양성분 계산기 로직 개선 및 결과 저장 기능 구현
- 영양성분 계산 알고리즘 정확도 향상
- 계산된 영양성분 데이터 SwiftData 저장 구현
- 사용자별 영양성분 계산 결과 히스토리 기능 추가
- 저장된 데이터 조회 및 관리 인터페이스 구현

+++ MeokPT/Sources/MyPage/Feature/BodyInfoInputFeature.swift
@@ -82,9 +82,9 @@ struct BodyInfoInputFeature: Reducer {
             case let .saveButtonTapped(context):
                  print("데이터 저장 시작")
                  do {
-                     let heightValue = Double(state.height) ?? 0
+                     let heightValue = Int(state.height) ?? 0
                      let ageValue = Int(state.age) ?? 0
-                     let weightValue = Double(state.weight) ?? 0
+                     let weightValue = Int(state.weight) ?? 0
                      let genderRawValue = state.selectedGender.rawValue
                      let goalRawValue = state.selectedGoal.rawValue
                      let activityLevelRawValue = state.selectedActivityLevel.rawValue
